flytter datakilder som kun skal hentes en gang per oppgave til egen datarute
