+++ app/hooks/useTypedRouteLoaderData.tsx
@@ -1,10 +1,12 @@
 import { useRouteLoaderData } from "@remix-run/react";
 import type { loader as rootLoader } from "~/root";
-import type { loader as oppgaveLoader } from "~/routes/saksbehandling.oppgave.$oppgaveId";
+import type { loader as oppgaveStatiskLoader } from "~/routes/saksbehandling.oppgave.$oppgaveId";
+import type { loader as oppgaveLoader } from "~/routes/saksbehandling.oppgave.$oppgaveId.steg";
 
 type Loaders = {
   root: typeof rootLoader;
-  "routes/saksbehandling.oppgave.$oppgaveId": typeof oppgaveLoader;
+  "routes/saksbehandling.oppgave.$oppgaveId": typeof oppgaveStatiskLoader;
+  "routes/saksbehandling.oppgave.$oppgaveId.steg": typeof oppgaveLoader;
 };
 
 export function useTypedRouteLoaderData<T extends keyof Loaders>(route: T) {
