+++ app/routes/saksbehandling.oppgave.$oppgaveId.tsx
@@ -3,40 +3,25 @@ import { json } from "@remix-run/node";
 import { Outlet, useLoaderData } from "@remix-run/react";
 import invariant from "tiny-invariant";
 import { hentOppgave } from "~/models/oppgave.server";
-
-import styles from "~/route-styles/behandle.module.css";
-import type { IJournalpost } from "~/models/SAF.server";
-import { hentJournalpost } from "~/models/SAF.server";
 import { getSession } from "~/models/auth.server";
-import { hentArbeidsforhold } from "~/models/arbeidsforhold.server";
+import styles from "~/route-styles/behandle.module.css";
 
 export async function loader({ params, request }: LoaderFunctionArgs) {
   invariant(params.oppgaveId, "params.oppgaveId er påkrevd");
-
   const session = await getSession(request);
   const oppgave = await hentOppgave(params.oppgaveId, session);
-  const arbeidsforhold = await hentArbeidsforhold(oppgave.person);
 
-  const journalposter: IJournalpost[] = [];
-
-  for (const journalpostId of oppgave.journalposter) {
-    const data = await hentJournalpost(request, journalpostId);
-    journalposter.push(data);
-  }
-
-  return json({ oppgave, journalposter, arbeidsforhold });
+  return json({ oppgave });
 }
 
 export default function PersonBehandle() {
   const { oppgave } = useLoaderData<typeof loader>();
   const forslagTilVedtak = oppgave.steg.find((steg) => steg.id === "Forslag til vedtak");
-
   return (
     <>
       {forslagTilVedtak?.tilstand === "Utført" && oppgave.tilstand === "TilBehandling" && (
         <div className={styles.innstiltBanner}>To-trinns kontroll</div>
       )}
-
       <div className={styles.container}>
         <Outlet />
       </div>
